+++ nens_auth_client/tests/conftest.py
@@ -5,6 +5,7 @@ import time
 import requests_mock
 import json
 import os
+from nens_auth_client.views import REDIRECT_SESSION_KEY
 
 DATA_PATH = os.path.join(os.path.dirname(__file__), "data")
 
@@ -87,10 +88,13 @@ def auth_req_generator(rf, mocker, rq_mocker, jwks, settings):
         settings.NENS_AUTH_AUTO_CREATE_SOCIAL_USER = False
 
         # Create the request
-        request = rf.get("/authorize?code={}&state={}".format(code, state))
+        request = rf.get(
+            "http://testserver/authorize?code={}&state={}".format(code, state)
+        )
         request.session = {
             "_cognito_authlib_state_": state,
             "_cognito_authlib_nonce_": nonce,
+            REDIRECT_SESSION_KEY: "http://testserver/success",
         }
         return request