+++ CHANGES.rst
@@ -16,3 +16,8 @@ Changelog of nens-auth-client
 - Added unittests for login and authorize views.
 
 - Added NENS_AUTH_ISSUER setting.
+
+- Parsing the next query parameter in the login view. If it is unsafe or not
+  provider, use ``NENS_AUTH_DEFAULT_SUCCESS_URL``.
+
+- If already logged in, the login view redirects to the success_url directly.